fix(scripts): run the main and dependency pkg scripts on the same terms
Use the same set of environment variables when running the main package
scripts and the dependencies package scripts.

close #319
