+++ lib/api/install.js
@@ -1,6 +1,5 @@
 'use strict'
 const path = require('path')
-const spawnSync = require('cross-spawn').sync
 
 const initCmd = require('./init_cmd')
 const installMultiple = require('../install_multiple')
@@ -8,6 +7,7 @@ const save = require('../save')
 const linkPeers = require('../install/link_peers')
 const runtimeError = require('../runtime_error')
 const getSaveType = require('../get_save_type')
+const runScript = require('../run_script')
 
 /*
  * Perform installation.
@@ -66,12 +66,12 @@ function install (packagesToInstall, opts) {
 
   function mainPostInstall () {
     const scripts = cmd.pkg.pkg && cmd.pkg.pkg.scripts || {}
-    if (scripts.postinstall) runScript('postinstall')
-    if (!isProductionInstall && scripts.prepublish) runScript('prepublish')
+    if (scripts.postinstall) npmRun('postinstall')
+    if (!isProductionInstall && scripts.prepublish) npmRun('prepublish')
     return
 
-    function runScript (scriptName) {
-      const result = spawnSync('npm', ['run', scriptName], {
+    function npmRun (scriptName) {
+      const result = runScript.sync('npm', ['run', scriptName], {
         cwd: path.dirname(cmd.pkg.path),
         stdio: 'inherit'
       })
